Revert some changes for API stability
