+++ libmamba/include/mamba/api/environment_yaml.hpp
@@ -8,14 +8,10 @@
 #define MAMBA_API_ENVIRONMENT_YAML_HPP
 
 #include <iosfwd>
-#include <map>
 #include <string>
-#include <vector>
 
-#include "mamba/fs/filesystem.hpp"
-
-// Include install.hpp for other_pkg_mgr_spec definition (needed by yaml_file_contents)
 #include "mamba/api/install.hpp"
+#include "mamba/fs/filesystem.hpp"
 
 namespace mamba
 {
@@ -29,19 +25,6 @@ namespace mamba
         bool ignore_channels = false;
         bool include_md5 = false;
     };
-
-    namespace detail
-    {
-        // yaml_file_contents is now defined here instead of install.hpp
-        struct yaml_file_contents
-        {
-            std::string name;
-            std::string prefix;
-            std::vector<std::string> dependencies, channels;
-            std::map<std::string, std::string> variables;
-            std::vector<other_pkg_mgr_spec> others_pkg_mgrs_specs;
-        };
-    }
 
     // Convert PrefixData to yaml_file_contents
     detail::yaml_file_contents prefix_to_yaml_contents(
